Add Go To Function definition
This only adds support for user-defined functions in the same file.

When the function definition cannot be found, I mirrored the experience
after C# and showed a message box.

Unlike C#, there is an extra case where there's multiple functions with
the same name and no way to know what function is being referenced.  In
this case, we arbitrarily choose the first one and display a message in
the status bar.